WIP pagination on results table
